+++ src/query/sql/src/planner/semantic/mod.rs
@@ -19,7 +19,9 @@ mod grouping_check;
 mod lowering;
 mod name_resolution;
 mod type_check;
+mod udf_rewriter;
 mod view_rewriter;
+mod virtual_column_rewriter;
 mod window_check;
 
 pub use aggregate_rewriter::AggregateRewriter;
@@ -36,7 +38,9 @@ pub use type_check::resolve_type_name;
 pub use type_check::resolve_type_name_by_str;
 pub use type_check::validate_function_arg;
 pub use type_check::TypeChecker;
+pub(crate) use udf_rewriter::UdfRewriter;
 pub use view_rewriter::ViewRewriter;
+pub(crate) use virtual_column_rewriter::VirtualColumnRewriter;
 pub use window_check::WindowChecker;
 
 pub(crate) const SUPPORTED_AGGREGATING_INDEX_FUNCTIONS: [&str; 5] =
