+++ src/query/storages/common/index/src/range_index.rs
@@ -84,7 +84,12 @@ impl RangeIndex {
             .column_refs()
             .into_iter()
             .map(|(name, ty)| {
-                if is_internal_column(&name) || is_stream_column(&name) {
+                // inernal column and stream column are not actual columns
+                // variant type may be virtual columns that are not included in leaf columns
+                if is_internal_column(&name)
+                    || is_stream_column(&name)
+                    || ty.remove_nullable() == DataType::Variant
+                {
                     return Ok((name, Domain::full(&ty)));
                 }
